Updated Steel Wind Strike

+++ Spells/2nd_Level/Gust_of_Wind/Gust_of_Wind.js
@@ -1,10 +1,5 @@
 const MACRONAME = "Gust_of_Wind.0.1.js"
 /*****************************************************************************************
- * Tasks for this macro
- *  1. Place a tile containing a VFX to mark the difficult terrain
- *  2. Spit out a message to chat describing the effect
- *  3. Remove the tile at the end of this actor's next turn
- * 
  * 05/31/22 0.1 Creation of Macro
  *****************************************************************************************/
 const MACRO = MACRONAME.split(".")[0]     // Trim of the version number and extension
@@ -52,8 +47,8 @@ async function doOff() {
         jez.log(`Deleted Tile ${TILE_ID}`)
     } 
     else if (args[1] === "Effect") {
-        let existingEffect = await tToken.actor.effects.find(i => i.data.label === EFFECT);
-        await existingEffect.delete()
+        let existingEffect = await aToken.actor.effects.find(i => i.data.label === args[2]);
+        if (existingEffect) await existingEffect.delete()
     } else {
         msg = `Some bad logic happened in ${MACRO}. Args[1] = ${args[1]}. Please tell Joe the tale.`
         ui.notifications.error(msg)
@@ -89,38 +84,21 @@ async function doOnUse() {
     jez.log("msg", msg);
     // ---------------------------------------------------------------------------------------
     // Add an effect to the active token that expires at the end of its next turn. 
-    // BUG: For some reason the special Duration code in this effect is tossing an error
-    //   times-up.js:31 times-up |  Could not process combat update  TypeError: Cannot read properties of undefined (reading 'startsWith')
-    //   at combatUpdate.js:96
-    //   at Map.filter (collection.mjs:66)
-    //   at Combat.handlePreUpdateCombat (combatUpdate.js:91)
-    //   at Combat.preUpdateCombat (combatUpdate.js:141)
-    //   at Wrapper.🎁call_wrapper [as call_wrapper] (libWrapper-wrapper.js:620)
-    //   at Combat.processOverTime (utils.js:794)
-    //   at async ClientDatabaseBackend._preUpdateDocumentArray (foundry.js:10203)
-    //   at async ClientDatabaseBackend._updateDocuments (foundry.js:10122)
-    //   at async Function.updateDocuments (document.mjs:373)
-    //   at async Combat.update (document.mjs:456)
-    //   at async MonksCombatTracker._onCombatControl (foundry.js:58779)
-    // I'm leaving it as I think it should be and will research the issue when I have time and better 
-    // connectivity.
     //
-    let gameRound = game.combat ? game.combat.round : 0;
-    let specialDuration = ["turnEndSource"]
     let effectData = {
         label: aItem.name,
         icon: aItem.img,
         origin: aToken.uuid,
         disabled: false,
-        duration: { rounds: 2, startRound: gameRound },
-        flags: { dae: { itemData: aItem, specialDuration: specialDuration } },
+        flags: { dae: { itemData: aItem } },
         changes: [
-            { key: `macro.itemMacro`, mode: jez.CUSTOM, value: `TILE ${fetchedTile.id}`, priority: 20 },
+            { key: `macro.itemMacro`, mode: jez.CUSTOM, value: `Tile ${fetchedTile.id}`, priority: 20 },
         ]
     };
     await MidiQOL.socket().executeAsGM("createEffects", { actorUuid: aActor.uuid, effects: [effectData] });
-    msg = `The ice storm leaves behind an area of difficult terrain until the end of 
-    <b>${aToken.name}</b>'s next turn.`
+    msg = `Strong wind blasts from <b>${aToken.name}</b>. Eash token that starts its turn in the 
+    gusts must make a DC${jez.getSpellDC(aToken)} STR Save or be moved 15 feet.  
+    Moving into the wind costs double.`
     postResults(msg)
     // ---------------------------------------------------------------------------------------
     // Modify the concentrating effect to trigger removal of the associated effect
@@ -144,7 +122,7 @@ async function placeTileVFX(TEMPLATE_ID, vfxFile, tilesWide, tilesHigh) {
     canvas.templates.get(TEMPLATE_ID).document.delete();
     // Place the tile with an embedded VFX
     let tileProps = {
-        x: template.center.x /*- GRID_SIZE*tilesWide/2*/,   // X coordinate is center of the template
+        x: template.center.x /*- GRID_SIZE*tilesWide/2*/,   // X coordinate is poorly understood
         y: template.center.y - GRID_SIZE * tilesHigh / 2,   // Y coordinate is center of the template
         img: vfxFile,
         width: GRID_SIZE * tilesWide,   // VFX should occupy 2 tiles across
@@ -169,7 +147,7 @@ async function modConcentratingEffect(tToken, label) {
     // Define the desired modification to existing effect. In this case, a world macro that will be
     // given arguments: VFX_Name and Token.id for all affected tokens
     //    
-    effect.data.changes.push({ key: `macro.itemMacro`, mode: jez.CUSTOM, value: `Effect ${label}`, priority: 20 })
+    effect.data.changes.push({ key: `macro.itemMacro`, mode: jez.CUSTOM, value: `Effect '${label}'`, priority: 20 })
     jez.log(`effect.data.changes`, effect.data.changes)
     //----------------------------------------------------------------------------------------------
     // Apply the modification to existing effect
